Adjusted server.js so that recursion loops should not happen

+++ server.js
@@ -1,15 +1,15 @@
 // imports
 const express = require('express');
 const axios = require('axios');
-const fs = require("fs");
+// const fs = require("fs");
 const { spawn } = require("child_process");
 const { updateEnv, log } = require("./common");
-require('dotenv').config(); 
+const { ref } = require('process');
 
+require('dotenv').config(); 
 const ENV_FILE = './.env';
-let ACCESS_TOKEN = '';
-let REFRESH_TOKEN = '';
-let EXPIRES_IN = '';
+
+let ACCESS_TOKEN, REFRESH_TOKEN, EXPIRES_IN;
 
 const app = express();
 const PORT = 3000;
@@ -17,26 +17,45 @@ const PORT = 3000;
 log(`server.js: start of server`)
 
 //get the token from twitch
-async function FetchToken({ code = null } = {}) {
+async function FetchToken({ code = null } = {}, refresh = false) {
   try{
-    if (!code) throw err;
-    const response = await axios.post('https://id.twitch.tv/oauth2/token', null, {
-      params: {
-        client_id: process.env.CLIENT_ID,
-        client_secret: process.env.CLIENT_SECRET,
-        code: code,
-        grant_type: 'authorization_code',
-        redirect_uri: process.env.REDIRECT_URI
-      }
-    });
+    let response;
+
+    if (refresh == true){
+      let refreshToken = REFRESH_TOKEN;
+      if (!refreshToken) throw new Error("No refresh token available");
+
+      response = await axios.post(
+        "https://id.twitch.tv/oauth2/token",
+        null,
+        {params: {
+          client_id: process.env.CLIENT_ID,
+          client_secret: process.env.CLIENT_SECRET,
+          grant_type: "refresh_token",
+          refresh_token: refreshToken,
+        },}
+      );
+
+    } else{
+      if (!code) throw new Error("No code provided");
+      response = await axios.post(
+        'https://id.twitch.tv/oauth2/token',
+        null,
+        {params: {
+          client_id: process.env.CLIENT_ID,
+          client_secret: process.env.CLIENT_SECRET,
+          code: code,
+          grant_type: 'authorization_code',
+          redirect_uri: process.env.REDIRECT_URI
+        }}
+      );
+    }
     
     const { access_token, refresh_token, expires_in } = response.data;
 
     ACCESS_TOKEN = access_token;
     REFRESH_TOKEN = refresh_token;
     EXPIRES_IN = expires_in;
-
-    CountDown(expires_in);
     // startCountdown(125);    //test
   
   } catch (err) {
@@ -45,73 +64,46 @@ async function FetchToken({ code = null } = {}) {
   }
 }
 
-async function RefreshAccessToken({ code = null } = {}) {
+
+async function Twitch_Auth({code = null} = {}){
   try{
-    let refreshToken = REFRESH_TOKEN;
-    if (!refreshToken) throw new Error("No refresh token available");
 
-    const response = await axios.post(
-      "https://id.twitch.tv/oauth2/token",
-      null,
-      {
-        params: {
-          client_id: process.env.CLIENT_ID,
-          client_secret: process.env.CLIENT_SECRET,
-          grant_type: "refresh_token",
-          refresh_token: refreshToken,
-        },
-      }
-    );
+    if (code) {
+      await FetchToken({ code });
+    }
 
-    const { access_token, refresh_token, expires_in } = response.data;
-    ACCESS_TOKEN = access_token;
-    REFRESH_TOKEN = refresh_token;
-    EXPIRES_IN = expires_in;
+    tmiProcess = null;
 
-    log(`server.js: refreshed the tokens`);
+    while(true){
+      
+      if (tmiProcess) {tmiProcess.kill();}
+      tmiProcess = spawn(
+        "node",
+        ["./twitch-bot.js", process.env.BOT, ACCESS_TOKEN],
+        {stdio: "inherit"}
+      );
+      
+      const refreshTime = (EXPIRES_IN - 120) * 1000;
+      await new Promise(resolve => setTimeout(resolve, refreshTime));
 
-    CountDown(expires_in);
-    // startCountdown(125);    //test
-
-  } catch (err) {
-    log(`server.js: RefreshAccessToken funct error: ${err.response?.data || err.message}`);
+      await FetchToken({}, refresh = true);
+    }
+  }
+  catch(err){
+    log(`server.js: Twitch_Auth error: ${err.response?.data || err.message}`);
     throw err;
   }
-}
 
-async function CountDown(expires_in) {
-  try{
-    
-    log(`server.js | Bot Name: ${process.env.BOT}`)
-    log(`server.js | ${ACCESS_TOKEN}`)
-
-    tmiProcess = spawn("node",
-      ["./twitch-bot.js", process.env.BOT, ACCESS_TOKEN],
-      {stdio: "inherit",});
-    
-    const refreshTime = (expires_in - 120) * 1000;
-    await new Promise(resolve => setTimeout(resolve, refreshTime));
-
-    if (tmiProcess) {
-      tmiProcess.kill();
-      tmiProcess = null;
-    }
-
-    await RefreshAccessToken();
-
-    } catch (err) {
-      log(`server.js: CountDown error: ${err.response?.data || err.message}`);
-      throw err;
-    }
 }
 
 
+// Server Response to Twitch
 app.get("/twitch-auth", async (req, res) => {
   try {
     const code = req.query.code;
     if (!code) return res.send("No code provided!");
 
-    await FetchToken({ code }); // just call the function
+    Twitch_Auth({ code }); // just call the function
     res.send("Token received, check console!");
 
   } catch {
